refactor: Improve product item pricing
